requirements.txt
arxiv==2.2.0
beautifulsoup4==4.13.4
dashscope==1.24.0
httpx==0.28.1
requests==2.32.4
pillow==11.3.0
python-dotenv==1.1.1
PyMuPDF==1.26.3
tqdm==4.67.1
